requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.18
transformers>=4.30.0
torch>=2.0.0
python-dotenv>=1.0.0
gunicorn>=20.1.0
psycopg2-binary>=2.9.6 